Added fixed LED target
Also break backwards compatibility a bit: The original firmware allows a
LED count of at most 254. However the RAM required by this LED count in
the original firmware exceeds the available RAM of the ATtiny85-20SU by
quite a lot. If I am not mistaken at least. We limit ourself to 128 LEDs.